requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
pydantic-settings[dotenv]
alembic
pyyaml